2022-12-30 10:38 PM

+++ DataStructures/lab/lab2/DebugPractice/DebugExercise1.java
@@ -5,7 +5,7 @@ package DataStructures.lab.lab2.DebugPractice;
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+        float quotient = (float)top / bottom;
         int result = Math.round(quotient);
         return result;
     }
